Just go back to search component being defined in Header.tsx for less diffs
