--- plugins/linear-genome-view/src/LinearGenomeView/components/Search.tsx
@@ -1,63 +1,0 @@
-import React, { useRef, useState } from 'react'
-import { getSession } from '@jbrowse/core/util'
-import { observer } from 'mobx-react'
-import SearchIcon from '@material-ui/icons/Search'
-import TextField from '@material-ui/core/TextField'
-import { useTheme } from '@material-ui/core/styles'
-import { fade } from '@material-ui/core/styles/colorManipulator'
-import { LinearGenomeViewModel } from '..'
-
-const WIDGET_HEIGHT = 32
-const SPACING = 7
-
-export default observer(({ model }: { model: LinearGenomeViewModel }) => {
-  const [value, setValue] = useState<string | undefined>()
-  const inputRef = useRef<HTMLInputElement>(null)
-  const theme = useTheme()
-  const { coarseVisibleLocStrings: visibleLocStrings } = model
-  const session = getSession(model)
-
-  function navTo(locString: string) {
-    try {
-      model.navToLocString(locString)
-    } catch (e) {
-      console.warn(e)
-      session.notify(`${e}`, 'warning')
-    }
-  }
-
-  return (
-    <form
-      data-testid="search-form"
-      onSubmit={event => {
-        event.preventDefault()
-        inputRef && inputRef.current && inputRef.current.blur()
-        value && navTo(value)
-
-        // need to manually call the action of blur here
-        setValue(undefined)
-      }}
-    >
-      <TextField
-        inputRef={inputRef}
-        onFocus={() => setValue(visibleLocStrings)}
-        onBlur={() => setValue(undefined)}
-        onChange={event => setValue(event.target.value)}
-        variant="outlined"
-        value={value === undefined ? visibleLocStrings : value}
-        style={{ margin: SPACING, marginLeft: SPACING * 3 }}
-        inputProps={{
-          'data-testid': 'search-input',
-        }}
-        // eslint-disable-next-line react/jsx-no-duplicate-props
-        InputProps={{
-          startAdornment: <SearchIcon />,
-          style: {
-            background: fade(theme.palette.background.paper, 0.8),
-            height: WIDGET_HEIGHT,
-          },
-        }}
-      />
-    </form>
-  )
-})
